Repo: Move the last source out of the root and finish the ESM migration
The root held three source files, one of which was a near-duplicate of a client
already in lib/. Moving it there would have collided with its own copy, so the two
are merged instead: lib/OpenDotaAPI.mjs is the survivor, and the handler it served
is now an ESM handler in handlers/ rather than a CommonJS body reached through an
adapter that existed only to give it a named export.

Merging meant picking one error convention. The two disagreed: one returned [] on
any failure, the other threw. The [] only bought resilience for getRecentMatches --
for getPlayer and getMatch it flowed on and crashed a few frames later at a property
access, so the run failed either way, just without naming a cause. The client now
throws, and the one call site that can recover from a failure says so and logs which
player it skipped.

Converting the last CommonJS also retires the createRequire bundling banner: esbuild
kept the require calls in those files and rewrote them to a shim that throws in ESM
output. Nothing bundled requires anything now, so the banner goes with them.
