--- OpenDotaAPI.js
@@ -1,125 +1,0 @@
-"use strict";
-
-const querystring = require("querystring");
-
-// Import cache as a dynamic import since it's an ESM module
-let cache;
-(async () => {
-  cache = (await import("./lib/cache.mjs")).default;
-})();
-
-const SEVEN_DAYS = 7 * 24 * 60 * 60;
-
-function API() {
-  const prefix = "https://api.opendota.com/api/";
-  // Resolves [] rather than throwing on any failure: a poll that misses one player
-  // should not take down the whole run.
-  async function getRequest(url, params) {
-    params = params ? params : {};
-
-    const requestUrl = `${prefix}${url}?${querystring.stringify(params)}`;
-
-    let body;
-    try {
-      const response = await fetch(requestUrl, {
-        signal: AbortSignal.timeout(10000),
-      });
-
-      if (!response.ok) {
-        console.error(`request error ${url}: HTTP ${response.status}`);
-        return [];
-      }
-
-      body = await response.text();
-    } catch (err) {
-      console.error(`request error ${url}:`);
-      console.error(err);
-      if (err.name === "TimeoutError") {
-        console.error(`Request timed out after 10 seconds for ${url}`);
-      }
-      return [];
-    }
-
-    try {
-      return JSON.parse(body);
-    } catch (parseError) {
-      console.error(
-        `Failed to parse JSON response: ${parseError.message}. Response preview: ${body.substring(0, 200)}`,
-      );
-      return [];
-    }
-  }
-
-  async function getMatch(matchID) {
-    const cacheNamespace = "opendota-matches";
-    const cacheKey = `match:${matchID}`;
-
-    // Check cache first
-    if (cache) {
-      const cachedMatch = await cache.get(cacheNamespace, cacheKey);
-      if (cachedMatch) {
-        console.log(`Cache hit for match ${matchID}`);
-        return JSON.parse(cachedMatch);
-      }
-      console.log(`Cache miss for match ${matchID}`);
-    }
-
-    // Fetch from API
-    const match = await getRequest(`matches/${matchID}`);
-
-    if (cache && match?.od_data?.has_parsed) {
-      await cache.set(
-        cacheNamespace,
-        cacheKey,
-        JSON.stringify(match),
-        SEVEN_DAYS,
-      );
-      console.log(`Cached parsed match ${matchID}`);
-    }
-
-    return match;
-  }
-
-  function getRecentMatches(steamID) {
-    return getRequest(`players/${steamID}/matches`, {
-      limit: 10,
-      significant: 0,
-    });
-  }
-
-  function getPlayer(steamID) {
-    return getRequest(`players/${steamID}`);
-  }
-
-  function getHeroItemPopularity(heroID) {
-    return getRequest(`heroes/${heroID}/itemPopularity`);
-  }
-
-  async function requestParse(matchID) {
-    try {
-      const response = await fetch(`${prefix}request/${matchID}`, {
-        method: "POST",
-      });
-
-      if (!response.ok) {
-        const responseText = await response.text();
-        console.error(
-          `Parse request failed for match ${matchID}: ${response.status} - ${responseText}`,
-        );
-      }
-
-      return response;
-    } catch (error) {
-      console.error(`Parse request error for match ${matchID}:`, error);
-      throw error;
-    }
-  }
-
-  this.getMatch = getMatch.bind(this);
-  this.getRecentMatches = getRecentMatches.bind(this);
-  this.getPlayer = getPlayer.bind(this);
-  this.getHeroItemPopularity = getHeroItemPopularity.bind(this);
-  this.requestParse = requestParse.bind(this);
-}
-
-module.exports = new API();
