+++ lib/Discord.mjs
@@ -1,22 +1,20 @@
-"use strict";
+export default class Discord {
+  #channels = {};
+  #userAgent = "";
+  #apikey = "";
 
-module.exports = function () {
-  let channels = {};
-  let userAgent = "";
-  let apikey = "";
-
-  function init(data) {
-    channels = data.channels;
-    userAgent = data.userAgent;
-    apikey = data.apikey;
+  init(data) {
+    this.#channels = data.channels;
+    this.#userAgent = data.userAgent;
+    this.#apikey = data.apikey;
   }
 
-  async function sendEmbed(embed, channel, components) {
+  async sendEmbed(embed, channel, components) {
     if (!channel) {
       channel = "default";
     }
 
-    if (!(channel in channels)) {
+    if (!(channel in this.#channels)) {
       console.error(`Unknown channel: ${channel}`);
       throw new Error(`Unknown channel: ${channel}`);
     }
@@ -34,12 +32,12 @@ module.exports = function () {
     let err;
     try {
       response = await fetch(
-        `https://discordapp.com/api/channels/${channels[channel]}/messages`,
+        `https://discordapp.com/api/channels/${this.#channels[channel]}/messages`,
         {
           method: "POST",
           headers: {
-            "User-Agent": userAgent,
-            Authorization: `Bot ${apikey}`,
+            "User-Agent": this.#userAgent,
+            Authorization: `Bot ${this.#apikey}`,
             "Content-Type": "application/json",
           },
           body: JSON.stringify(payload),
@@ -64,7 +62,7 @@ module.exports = function () {
     return { error: hadError };
   }
 
-  async function sendInteractionResponse(
+  async sendInteractionResponse(
     applicationID,
     interactionToken,
     payload,
@@ -91,8 +89,4 @@ module.exports = function () {
 
     return response;
   }
-
-  this.init = init.bind(this);
-  this.sendEmbed = sendEmbed.bind(this);
-  this.sendInteractionResponse = sendInteractionResponse.bind(this);
-};
+}
