+++ lib/OpenDotaAPI.mjs
@@ -1,79 +1,59 @@
-import querystring from 'querystring';
-
 const SEVEN_DAYS = 7 * 24 * 60 * 60;
-const API_PREFIX = 'https://api.opendota.com/api/';
-const CACHE_NAMESPACE = 'opendota-matches';
+const API_PREFIX = "https://api.opendota.com/api/";
+const CACHE_NAMESPACE = "opendota-matches";
+const TIMEOUT_MS = 10000;
 
-/**
- * OpenDota API client with optional caching support.
- * Use constructor to inject a cache implementation or pass null for no caching.
- */
-class OpenDotaAPI {
+export default class OpenDotaAPI {
   #cache;
 
-  /**
-   * @param {Object|null} cache - Cache implementation with get/set methods, or null to disable caching
-   */
+  /** @param {{get, set}|null} cache - omit to disable caching */
   constructor(cache = null) {
     this.#cache = cache;
   }
 
-  /**
-   * Make a GET request to the OpenDota API
-   * @private
-   */
-  async #getRequest(url, params = {}) {
-    const requestUrl = `${API_PREFIX}${url}?${querystring.stringify(params)}`;
+  async #getRequest(path, params) {
+    const query = params ? `?${new URLSearchParams(params)}` : "";
 
-    try {
-      const response = await fetch(requestUrl);
-      if (!response.ok) {
-        console.error(`OpenDota API error ${url}: ${response.status}`);
-      }
-      return await response.json();
-    } catch (err) {
-      console.error(`Request error ${url}:`, err);
-      throw err;
+    const response = await fetch(`${API_PREFIX}${path}${query}`, {
+      signal: AbortSignal.timeout(TIMEOUT_MS),
+    });
+
+    if (!response.ok) {
+      throw new Error(`OpenDota ${path}: HTTP ${response.status}`);
     }
+
+    return response.json();
   }
 
-  /**
-   * Get match data by match ID with optional caching
-   * @param {string|number} matchID - Match ID
-   * @returns {Promise<Object>} Match data
-   */
   async getMatch(matchID) {
     const cacheKey = `match:${matchID}`;
 
-    // Check cache first
     if (this.#cache) {
       const cachedMatch = await this.#cache.get(CACHE_NAMESPACE, cacheKey);
       if (cachedMatch) {
         console.log(`Cache hit for match ${matchID}`);
-        return typeof cachedMatch === 'string'
-          ? JSON.parse(cachedMatch)
-          : cachedMatch;
+        return JSON.parse(cachedMatch);
       }
       console.log(`Cache miss for match ${matchID}`);
     }
 
-    // Fetch from API
     const match = await this.#getRequest(`matches/${matchID}`);
 
-    // Cache parsed matches
+    // Only parsed matches are worth keeping: an unparsed one is missing the fields
+    // the analyst reads, and OpenDota fills them in later.
     if (this.#cache && match?.od_data?.has_parsed) {
-      await this.#cache.set(CACHE_NAMESPACE, cacheKey, JSON.stringify(match), SEVEN_DAYS);
+      await this.#cache.set(
+        CACHE_NAMESPACE,
+        cacheKey,
+        JSON.stringify(match),
+        SEVEN_DAYS,
+      );
       console.log(`Cached parsed match ${matchID}`);
     }
 
     return match;
   }
 
-  /**
-   * Get recent matches for a player
-   * @param {string|number} steamID - Steam account ID
-   * @returns {Promise<Array>} Recent matches
-   */
   async getRecentMatches(steamID) {
     return this.#getRequest(`players/${steamID}/matches`, {
       limit: 10,
@@ -81,49 +61,26 @@ class OpenDotaAPI {
     });
   }
 
-  /**
-   * Get player data
-   * @param {string|number} steamID - Steam account ID
-   * @returns {Promise<Object>} Player data
-   */
   async getPlayer(steamID) {
     return this.#getRequest(`players/${steamID}`);
   }
 
-  /**
-   * Get hero item popularity statistics
-   * @param {string|number} heroID - Hero ID
-   * @returns {Promise<Object>} Item popularity data
-   */
   async getHeroItemPopularity(heroID) {
     return this.#getRequest(`heroes/${heroID}/itemPopularity`);
   }
 
-  /**
-   * Request match parsing
-   * @param {string|number} matchID - Match ID
-   * @returns {Promise<Response>} Fetch response
-   */
   async requestParse(matchID) {
-    try {
-      const response = await fetch(`${API_PREFIX}request/${matchID}`, {
-        method: 'POST',
-      });
+    const response = await fetch(`${API_PREFIX}request/${matchID}`, {
+      method: "POST",
+    });
 
-      if (!response.ok) {
-        const responseText = await response.text();
-        console.error(
-          `Parse request failed for match ${matchID}: ${response.status} - ${responseText}`,
-        );
-      }
-
-      return response;
-    } catch (error) {
-      console.error(`Parse request error for match ${matchID}:`, error);
-      throw error;
+    if (!response.ok) {
+      const body = await response.text();
+      console.error(
+        `Parse request failed for match ${matchID}: ${response.status} - ${body}`,
+      );
     }
+
+    return response;
   }
 }
-
-// Export only the class
-export default OpenDotaAPI;
