+++ lib/cache.mjs
@@ -6,7 +6,7 @@ import {
 } from "@aws-sdk/lib-dynamodb";
 import { gzip, gunzip } from "zlib";
 import { promisify } from "util";
-import tables from "./tables.js";
+import tables from "./tables.mjs";
 
 const gzipAsync = promisify(gzip);
 const gunzipAsync = promisify(gunzip);
