+++ lib/utils.mjs
@@ -1,11 +1,9 @@
-"use strict";
-
-const { DynamoDBClient } = require("@aws-sdk/client-dynamodb");
-const {
+import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
+import {
   DynamoDBDocumentClient,
   GetCommand,
   PutCommand,
-} = require("@aws-sdk/lib-dynamodb");
+} from "@aws-sdk/lib-dynamodb";
 
 const client = new DynamoDBClient({});
 const docClient = DynamoDBDocumentClient.from(client);
@@ -16,7 +14,7 @@ const USER_AGENT = "c0rbot/1.0 (+https://github.com/coreymaher/c0rbot)";
 
 // Resolves the response body, or `undefined` on any failure -- callers treat a falsy
 // result as a transient error and skip the run.
-module.exports.simpleGet = async (url, { qs, headers, timeout } = {}) => {
+export async function simpleGet(url, { qs, headers, timeout } = {}) {
   const target = qs ? `${url}?${new URLSearchParams(qs)}` : url;
 
   try {
@@ -36,9 +34,9 @@ module.exports.simpleGet = async (url, { qs, headers, timeout } = {}) => {
     console.error(err);
     return undefined;
   }
-};
+}
 
-module.exports.loadFeedData = async (key) => {
+export async function loadFeedData(key) {
   try {
     const params = {
       TableName: process.env.table,
@@ -53,9 +51,9 @@ module.exports.loadFeedData = async (key) => {
     console.error({ TableName: process.env.table, Key: { name: key } });
     return {};
   }
-};
+}
 
-module.exports.updateFeedData = async (key, feedData) => {
+export async function updateFeedData(key, feedData) {
   try {
     const params = {
       TableName: process.env.table,
@@ -75,4 +73,4 @@ module.exports.updateFeedData = async (key, feedData) => {
       Item: { name: key, feed_data: feedData, updated_at: Date.now() },
     });
   }
-};
+}
